made CVS compile under linux again
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@57 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Pixes/pix_aging.cpp
@@ -250,7 +250,7 @@ void pix_aging :: colorMessCallback(void *data, t_floatarg state)
 }
 void pix_aging :: scratchMessCallback(void *data, t_floatarg state)
 {
-  GetMyClass(data)->scratchMess(state);  
+  GetMyClass(data)->scratchMess((int)state);  
 }
 void pix_aging :: dustMessCallback(void *data, t_floatarg state)
 {
